Moved ALL of the routing functionality for hitting the /city route into a new router handler -- requests at /api/city are now dealt with by a router.js script which is imported. Atm, all of the functionality is identical, it's just modularized.
